Move more of the extension construction to setupext.py
