+++ setupext.py
@@ -1,17 +1,23 @@
+import os
+import platform
 import contextlib
 import glob
-import os
 import shutil
 import subprocess
 import sys
+from sys import platform as _platform
 import tempfile
+from concurrent.futures import ThreadPoolExecutor as Pool
 
 from distutils import log
-from distutils.ccompiler import new_compiler
+from distutils.ccompiler import new_compiler, CCompiler
 from distutils.sysconfig import customize_compiler
 from distutils.errors import CompileError, LinkError
+from distutils.version import LooseVersion
 from pkg_resources import resource_filename
 from subprocess import Popen, PIPE
+from setuptools.command.build_ext import build_ext as _build_ext
+from setuptools.command.sdist import sdist as _sdist
 
 
 CCODE = """
@@ -124,12 +130,32 @@ def check_for_openmp():
     return using_openmp
 
 
-def check_for_pyembree():
+def check_for_pyembree(std_libs):
+    embree_libs = []
+    embree_aliases = {}
     try:
         fn = resource_filename("pyembree", "rtcore.pxd")
     except ImportError:
-        return None
-    return os.path.dirname(fn)
+        return embree_libs, embree_aliases
+
+    embree_prefix = os.path.abspath(read_embree_location())
+    embree_inc_dir = os.path.join(embree_prefix, 'include')
+    embree_lib_dir = os.path.join(embree_prefix, 'lib')
+    if in_conda_env():
+        conda_basedir = os.path.dirname(os.path.dirname(sys.executable))
+        embree_inc_dir.append(os.path.join(conda_basedir, 'include'))
+        embree_lib_dir.append(os.path.join(conda_basedir, 'lib'))
+
+    if _platform == "darwin":
+        embree_lib_name = "embree.2"
+    else:
+        embree_lib_name = "embree"
+
+    embree_aliases['EMBREE_INC_DIR'] = ['yt/utilities/lib/', embree_inc_dir]
+    embree_aliases['EMBREE_LIB_DIR'] = [embree_lib_dir]
+    embree_aliases['EMBREE_LIBS'] = std_libs + [embree_lib_name]
+    embree_libs += ['yt/utilities/lib/embree_mesh/*.pyx']
+    return embree_libs, embree_aliases
 
 def in_conda_env():
     return any(s in sys.version for s in ("Anaconda", "Continuum"))
@@ -206,3 +232,104 @@ def read_embree_location():
         shutil.rmtree(tmpdir)
 
     return rd
+
+def get_cpu_count():
+    if platform.system() != "Windows":
+        return os.cpu_count()
+    return 0
+
+def install_ccompiler():
+    def _compile(
+        self, sources, output_dir=None, macros=None, include_dirs=None,
+        debug=0, extra_preargs=None, extra_postargs=None, depends=None,
+    ):
+        """Function to monkey-patch distutils.ccompiler.CCompiler"""
+        macros, objects, extra_postargs, pp_opts, build = self._setup_compile(
+            output_dir, macros, include_dirs, sources, depends, extra_postargs
+        )
+        cc_args = self._get_cc_args(pp_opts, debug, extra_preargs)
+
+        for obj in objects:
+            try:
+                src, ext = build[obj]
+            except KeyError:
+                continue
+            self._compile(obj, src, ext, cc_args, extra_postargs, pp_opts)
+
+        # Return *all* object filenames, not just the ones we just built.
+        return objects
+
+    CCompiler.compile = _compile
+
+def create_build_ext(lib_exts, cythonize_aliases):
+    class build_ext(_build_ext):
+        # subclass setuptools extension builder to avoid importing cython and numpy
+        # at top level in setup.py. See http://stackoverflow.com/a/21621689/1382869
+        def finalize_options(self):
+            try:
+                import cython
+                import numpy
+            except ImportError:
+                raise ImportError(
+    """Could not import cython or numpy. Building yt from source requires
+    cython and numpy to be installed. Please install these packages using
+    the appropriate package manager for your python environment.""")
+            if LooseVersion(cython.__version__) < LooseVersion('0.26.1'):
+                raise RuntimeError(
+    """Building yt from source requires Cython 0.26.1 or newer but
+    Cython %s is installed. Please update Cython using the appropriate
+    package manager for your python environment.""" %
+                    cython.__version__)
+            if LooseVersion(numpy.__version__) < LooseVersion('1.13.3'):
+                raise RuntimeError(
+    """Building yt from source requires NumPy 1.13.3 or newer but
+    NumPy %s is installed. Please update NumPy using the appropriate
+    package manager for your python environment.""" %
+                    numpy.__version__)
+            from Cython.Build import cythonize
+            # Override the list of extension modules
+            self.distribution.ext_modules[:] = cythonize(
+                lib_exts,
+                aliases = cythonize_aliases,
+                compiler_directives={'language_level': 2},
+                nthreads=get_cpu_count(),
+            )
+            _build_ext.finalize_options(self)
+            # Prevent numpy from thinking it is still in its setup process
+            # see http://stackoverflow.com/a/21621493/1382869
+            if isinstance(__builtins__, dict):
+                # sometimes this is a dict so we need to check for that
+                # https://docs.python.org/3/library/builtins.html
+                __builtins__["__NUMPY_SETUP__"] = False
+            else:
+                __builtins__.__NUMPY_SETUP__ = False
+            import numpy
+            self.include_dirs.append(numpy.get_include())
+
+        def build_extensions(self):
+            self.check_extensions_list(self.extensions)
+
+            ncpus = get_cpu_count()
+            if ncpus > 0:
+                with Pool(ncpus) as pool:
+                    pool.map(self.build_extension, self.extensions)
+            else:
+                super().build_extensions()
+
+
+    class sdist(_sdist):
+        # subclass setuptools source distribution builder to ensure cython
+        # generated C files are included in source distribution.
+        # See http://stackoverflow.com/a/18418524/1382869
+        def run(self):
+            # Make sure the compiled Cython files in the distribution are up-to-date
+            from Cython.Build import cythonize
+            cythonize(
+                lib_exts,
+                aliases = cythonize_aliases,
+                compiler_directives={'language_level': 2},
+                nthreads=get_cpu_count(),
+            )
+            _sdist.run(self)
+
+    return build_ext, sdist
